requirements.txt
django
psycopg2
djangorestframework
django_seed
